Move group endpoints into separate file
